Bug 30632: Fix report author display in list of saved reports
This patch removes whitespace in the template around the markup of
report author first name and surname. This prevents unwanted spaces in
the output.

To test, apply the patch and go to Reports -> Use saved.

The display of report author should be like this:

Leonard, Owen (123456)

Instead of Leonard , Owen (20375)

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>