Bug 30695: (QA follow-up) Catch moremember.tt too
This patch does the same treatment for moremember.tt where the same
summary block is includes. We also correct the tab order to be a little
more consistent too.

We should really move this to an include and refer to it from both
templates instead of this copy/paste code..

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>